backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3.1
drf-spectacular>=0.27.2
django-filter>=24.3
psycopg2-binary>=2.9.9
pytest>=8
pytest-django>=4.8
